feat(invoice): generate invoice from project materials with transactional service

+++ src/main/java/com/cabinetpro/lite/dto/InvoiceDto.java
@@ -3,11 +3,13 @@ package com.cabinetpro.lite.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data @AllArgsConstructor
 public class InvoiceDto {
     private Long id;
-    private double subtotal;
-    private double gst;
-    private double total;
+    private BigDecimal subtotal;
+    private BigDecimal gst;
+    private BigDecimal total;
     private String status; // DRAFT/PAID...
 }
