+++ src/main/java/com/cabinetpro/lite/dto/InvoiceGenerateRequestDto.java
@@ -2,8 +2,10 @@ package com.cabinetpro.lite.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InvoiceGenerateRequestDto {
     // اختیاری؛ اگر null باشد، 0.10
-    private Double gstRate;
+    private BigDecimal gstRate;
 }
